moved CreateAccount as Register component and created a RenderReg page to render page outside of appjs
