+++ client/src/components/Register/index.js
@@ -1,0 +1,178 @@
+import React, { Component } from "react";
+import "./style.css";
+
+const emailRegex = RegExp(
+  /^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/
+);
+
+const formValid = ({ formErrors, ...rest }) => {
+  let valid = true;
+
+  //validating form errors being empty
+  Object.values(formErrors).forEach((val) => {
+    val.length > 0 && (valid = false);
+  });
+
+  //validating form was filled
+  Object.values(rest).forEach((val) => {
+    val === null && (valid = false);
+  });
+
+  return valid;
+};
+
+class Register extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      firstName: null,
+      lastName: null,
+      email: null,
+      password: null,
+      formErrors: {
+        firstName: "",
+        lastName: "",
+        email: "",
+        password: "",
+      },
+    };
+  }
+
+  handleSubmit = (e) => {
+    e.preventDefault();
+
+    if (formValid(this.state)) {
+      console.log(`
+      ---SUBMITTING---
+      First Name: ${this.state.firstName}
+      Last Name: ${this.state.lastName}
+      Email: ${this.state.email}
+      Password: ${this.state.password}
+  `);
+    } else {
+      console.error("FORM INVALID - SEE ERROR MESSAGES");
+    }
+  };
+
+  handleChange = (e) => {
+    e.preventDefault();
+    const { name, value } = e.target;
+    let formErrors = this.state.formErrors;
+
+    switch (name) {
+      case "firstName":
+        formErrors.firstName =
+          value.length < 2 ? "minimum 2 characters required" : "";
+        break;
+      case "lastName":
+        formErrors.lastName =
+          value.length < 2 ? "minimum 2 characters required" : "";
+        break;
+      case "email":
+        formErrors.email = emailRegex.test(value)
+          ? ""
+          : "invalid email address";
+        break;
+      case "password":
+        formErrors.password =
+          value.length < 6 ? "minimum 6 characters required" : "";
+        break;
+      default:
+        break;
+    }
+
+    this.setState({ formErrors, [name]: value }, () => console.log(this.state));
+  };
+
+  render() {
+    const { formErrors } = this.state;
+    return (
+      <div className="wrapper">
+        <div className="form-wrapper">
+          <h1>Create Account</h1>
+          <form onSubmit={this.handleSubmit} noValidate>
+            {
+              //Input for First Name
+            }
+            <div className="firstName">
+              <label htmlFor="firstName"> First Name</label>
+              <input
+                className={formErrors.firstName.length > 0 ? "error" : null}
+                placeholder="First Name"
+                type="text"
+                name="firstName"
+                formNoValidate
+                onChange={this.handleChange}
+              />
+              {formErrors.firstName.length > 0 && (
+                <span className="errorMessage">{formErrors.firstName}</span>
+              )}
+            </div>
+            {
+              //Form input for Last Name
+            }
+            <div className="lastName">
+              <label htmlFor="lastName"> Last Name</label>
+              <input
+                className={formErrors.lastName.length > 0 ? "error" : null}
+                placeholder="Last Name"
+                type="text"
+                name="lastName"
+                formNoValidate
+                onChange={this.handleChange}
+              />
+              {formErrors.lastName.length > 0 && (
+                <span className="errorMessage">{formErrors.lastName}</span>
+              )}
+            </div>
+            {
+              //Form input for Email
+            }
+            <div className="email">
+              <label htmlFor="email"> Email</label>
+              <input
+                className={formErrors.email.length > 0 ? "error" : null}
+                placeholder="Email"
+                type="email"
+                name="email"
+                formNoValidate
+                onChange={this.handleChange}
+              />
+              {formErrors.email.length > 0 && (
+                <span className="errorMessage">{formErrors.email}</span>
+              )}
+            </div>
+            {
+              //Form input for Password
+            }
+            <div className="password">
+              <label htmlFor="password"> Password</label>
+              <input
+                className={formErrors.password.length > 0 ? "error" : null}
+                placeholder="Password"
+                type="password"
+                name="password"
+                formNoValidate
+                onChange={this.handleChange}
+              />
+              {formErrors.password.length > 0 && (
+                <span className="errorMessage">{formErrors.password}</span>
+              )}
+            </div>
+
+            {
+              //Create Account Button
+            }
+            <div className="createAccount">
+              <button type="submit"> Create Account </button>
+              <small>Already Have an Account?</small>
+            </div>
+          </form>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Register;
