+++ client/src/components/Register/style.css
@@ -1,0 +1,128 @@
+/* Colors */
+
+/*
+Brown: #322A26
+Light Olive: #95B46A
+Dark Moss: #709255
+*/
+
+body {
+    font-family: "Jost", sans-serif;
+}
+
+.wrapper {
+    height: 100vh;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: #322a26;
+}
+
+.form-wrapper {
+    width: 400px;
+    display: flex;
+    flex-direction: column;
+    padding: 20px 40px;
+    border-radius: 10px;
+    background-color: #95b46a;
+}
+
+form {
+    width: 100%;
+    display: flex;
+    flex-wrap: wrap;
+}
+
+h1 {
+    text-align: center;
+    width: 100%;
+    color: #322a26;
+    font-weight: 600;
+    font-family: "Jost", sans-serif;
+}
+
+label {
+    font-size: 0.8em;
+    margin-bottom: 0.25em;
+    color: #333138;
+    font-weight: 400;
+}
+
+input {
+    padding: 10px 10px;
+    border-radius: 5px;
+    outline: none;
+    border: 1px solid #709255;
+}
+
+input::placeholder {
+    font-size: 1.2em;
+    font-weight: lighter;
+    color: #709255;
+}
+
+input.error {
+    border: 1px solid red;
+}
+
+.errorMessage {
+    color: red;
+    font-size: 0.75em;
+    display: relative;
+}
+
+.firstName {
+    margin-right: 1%;
+}
+
+.lastName {
+    margin-left: 1%;
+}
+
+.firstName, .lastName, .email, .password {
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 15px;
+}
+
+.firstName, .lastName {
+    width: 49%;
+}
+
+.email, .password {
+    width: 100%;
+}
+
+.createAccount {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.createAccount button {
+    background-color: #709255;
+    color: #fff;
+    border: 2px solid #709255;
+    border-radius: 10px;
+    width: 100%;
+    margin-top: 1em;
+    padding: 8px 0px;
+    font-size: 1em;
+    font-weight: lighter;
+    letter-spacing: 1px;
+    margin-bottom: 0.25em;
+}
+
+.createAccount button:hover {
+    color: #322a26;
+    background-color: #709255;
+    border: 2px solid #95b46a;
+}
+
+.createAccount small {
+    color: #322a26;
+    font-weight: lighter;
+}
