+++ client/src/pages/Register.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Register from "../components/Register";
+
+function RenderReg() {
+  return <Register />;
+}
+
+export default RenderReg;
