+++ package.json
@@ -26,6 +26,7 @@
         "mdbootstrap": "^4.19.1",
         "mongoose": "^5.8.11",
         "passport": "^0.4.1",
-        "path": "^0.12.7"
+        "path": "^0.12.7",
+        "react-router-dom": "^5.2.0"
     }
 }
